Chnaged all IDs from String to Long

+++ src/main/java/com/cput/laundryecommercebookingsystem/domain/OrderItem.java
@@ -16,7 +16,7 @@ public class OrderItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_item_id", nullable = false)
-    private int orderItemId;
+    private Long orderItemId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id", nullable = false)
@@ -55,9 +55,7 @@ public class OrderItem {
     public void calculateSubtotal() {
         this.subtotal = this.quantity * this.unitPrice;
     }
-    public int getOrderItemId() {
-        return orderItemId;
-    }
+    public Long getOrderItemId() {return orderItemId;}
     public Order getOrder() {
         return order;
     }
@@ -114,13 +112,13 @@ public class OrderItem {
      */
     public static class Builder {
 
-        private int orderItemId;
+        private Long orderItemId;
         private Order order;
         private Product product;
         private int quantity;
         private double unitPrice;
 
-        public Builder setOrderItemId(int orderItemId) {
+        public Builder setOrderItemId(Long orderItemId) {
             this.orderItemId = orderItemId;
             return this;
         }
